extract PhysicsState, PhysicsSnapshot & PhysicsEvents
cargo fmt

changes to make rapier compile
