+++ src_testbed/lib.rs
@@ -23,7 +23,7 @@ extern crate log;
 
 pub use crate::engine::GraphicsManager;
 pub use crate::plugin::TestbedPlugin;
-pub use crate::testbed::{PhysicsState, Testbed};
+pub use crate::testbed::Testbed;
 
 #[cfg(all(feature = "dim2", feature = "other-backends"))]
 mod box2d_backend;
@@ -31,6 +31,7 @@ mod engine;
 #[cfg(feature = "other-backends")]
 mod nphysics_backend;
 pub mod objects;
+mod physics;
 #[cfg(all(feature = "dim3", feature = "other-backends"))]
 mod physx_backend;
 mod plugin;
